+++ packages/backend-auth/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @aws-amplify/backend-auth
 
+## 0.5.0-beta.9
+
+### Patch Changes
+
+- Updated dependencies [fe59a49]
+  - @aws-amplify/auth-construct-alpha@0.6.0-beta.9
+
+## 0.5.0-beta.8
+
+### Patch Changes
+
+- Updated dependencies [394b72e]
+  - @aws-amplify/auth-construct-alpha@0.6.0-beta.8
+
 ## 0.5.0-beta.7
 
 ### Patch Changes